[RHTAP-1069] Moved methods to correct files
