+++ pkg/utils/release/factory.go
@@ -4,12 +4,6 @@ import kubeCl "github.com/redhat-appstudio/e2e-tests/pkg/apis/kubernetes"
 
 // Contains all embebed interfaces for release operations.
 type ReleaseService interface {
-	// Interface for all component related operations.
-	ComponentsInterface
-
-	// Interface for all pipeline related operations.
-	PipelinesInterface
-
 	// Interface for all plan related operations.
 	PlansInterface
 
@@ -18,12 +12,6 @@ type ReleaseService interface {
 
 	// Interface for all sbom related operations.
 	SbomInterface
-
-	// Interface for all secret related operations.
-	SecretsInterface
-
-	// Interface for all snapshot related operations.
-	SnapshotsInterface
 
 	// Interface for all strategy related operations.
 	StrategiesInterface
